Move test folder into src for consistency with core
